tools/requirements.txt
rich>=13.0.0
requests>=2.31.0
psutil>=5.9.0
openai>=1.0.0
python-dotenv>=1.0.0
